fix bz5771054 info msg is application.json is missing
The intent of the bug was to inform users that they may have
tried to start their app from an incorrect working directory.

The new start command already prevents this by checking for
both a valid package.json, and that the dependencies field
contains "mojito". Adding this warning message and tests for
fuller compliance wrt application.json (which is technically
optional).